requirements.txt
Django==5.1.6
djangorestframework==3.15.2
pytest-django==4.10.0
pillow==11.1.0
psycopg==3.2.5
djangorestframework_simplejwt==5.5.0